+++ hitbox-redis/tests/integration_test.rs
@@ -5,7 +5,7 @@ use tokio::time::{sleep, Duration};
 
 #[actix_rt::test]
 async fn test_rw() -> Result<(), Error> {
-    let addr = RedisBuilder::single("redis://127.0.0.1:6379")
+    let addr = RedisBuilder::standalone("redis://127.0.0.1:6379")
         .await?
         .start();
     let message = Set {
@@ -29,7 +29,7 @@ async fn test_rw() -> Result<(), Error> {
 
 #[actix_rt::test]
 async fn test_set_expired() -> Result<(), Error> {
-    let addr = RedisBuilder::single("redis://127.0.0.1:6379")
+    let addr = RedisBuilder::standalone("redis://127.0.0.1:6379")
         .await?
         .start();
     let message = Set {
@@ -60,7 +60,7 @@ async fn test_set_expired() -> Result<(), Error> {
 
 #[actix_rt::test]
 async fn test_delete() -> Result<(), Error> {
-    let addr = RedisBuilder::single("redis://127.0.0.1:6379")
+    let addr = RedisBuilder::standalone("redis://127.0.0.1:6379")
         .await?
         .start();
     let message = Set {
@@ -95,7 +95,7 @@ async fn test_delete() -> Result<(), Error> {
 
 #[actix_rt::test]
 async fn test_lock() -> Result<(), Error> {
-    let addr = RedisBuilder::single("redis://127.0.0.1:6379")
+    let addr = RedisBuilder::standalone("redis://127.0.0.1:6379")
         .await?
         .start();
     let message = Lock {