+++ hitbox-redis/CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Add redis cluster support; Redis backend is split into single and cluster instances
+- Add redis cluster support; Redis backend is split into standalone and cluster instances
 
 ## [0.1.0] - 2021-05-29
 ### Added